CAmbianbdo el tipo del atributo

+++ Concepto.groovy
@@ -1,7 +1,7 @@
 class Concepto{
 	Integer cantidad
 	String descripcion 
-	Float importe
+	BigDecimal importe
 
 	String toString(){
 	  "Cantidad: ${cantidad}  Descripcion: ${descripcion} Importe: ${importe}"
